Address PR #62 review feedback (Feetech client cleanup)
- Drop redundant read_pos_vel_cur sync wrapper; the public method now does
  the sync read directly with per-motor fallback
- Rename _read_*_individual fallbacks to _read_*_per_motor_fallback so the
  intent is obvious at the call site
- Centralize POSITION_DIRECTION via _raw_to_rad / _rad_to_raw helpers so
  the sign inversion stops leaking into read/write logic
- Define MotorError + MotionTimeoutError; wait_for_motion_complete now
  raises on timeout instead of silently returning False
- Add MotorClient.waits_for_motion class flag; OrcaHand.wait_for_motion
  skips the motor lock when the underlying client doesn't actually block
- Strip the verbose AI-style comment from calibrate_offset
- Add explicit parens to the sign-flip in write_positions_sync
- Rename scripts/test.py -> scripts/stress_test.py and update docs
